fix(graphics): harden 2D/3D runtime and overhaul reference demos
Extend the runtime correctness-and-hot-path pass from the 3D graphics layer into the 2D graphics and physics runtime, add matching regression and fuzz coverage, and land a ground-up overhaul of the three reference games that exercise the hardened surface.

2D graphics and physics runtime:
- resolve broad-phase collisions through a sorted, de-duplicated candidate-pair scratch that grows geometrically (amortized O(1) appends) and falls back to the exhaustive O(n^2) pass on allocation failure, so every pair resolves at most once per step and in a deterministic order
- eject penetrating bodies along the true minimum-translation axis instead of an inflated intersection width, and add a distance-joint velocity-projection pass that cancels the phantom radial velocity the positional solver previously let accumulate under gravity
- make camera math exact and platform-stable with saturating int64 subtraction and an integer round(value*mul/div) fast path, keeping VM and native positions identical and avoiding both platform-varying long double and the __divti3 compiler-rt helper the native-link runtime must not depend on
- cache each sprite's transformed frame (flip, scale, pad, rotate, tint, alpha) keyed on every transform input plus the frame pointer, so static sprites stop re-running the whole pipeline and its allocations every frame
- reuse a per-scene draw-order scratch across frames so a static scene no longer pays a per-frame sequence allocation, and short-circuit subtree dirty marking
- clamp externally parsed JSON binding numbers to int64 before casting (a non-finite or out-of-range double cast is undefined behavior) and buffer bindings until the action object closes so name/type/bindings keys may appear in any order
- switch large downscales (>2x) to a premultiplied-alpha area filter so thumbnails and mip levels stop aliasing, blend the SpriteBatch region fast path so transparent frames composite instead of overwrite, fill scanlines through a scale-aware span primitive so HiDPI canvases stop striping, and interpolate canvas gradients per channel at full precision rather than through 0..100 percent quantization

3D runtime:
- weight angular joint constraints by a world-space inverse inertia tensor (rotate into the body's principal axes, scale, rotate back) solved with a 3x3 cofactor inverse, rather than by inverse mass
- reject Draco KHR_draco_mesh_compression payloads whose element counts are grossly disproportionate to the byte size or whose NORMALS transform stride disagrees with the declared components, bound the edgebreaker leftmost-corner walk, and expose an internal rt_gltf_draco_decode_probe entry point for direct testing and fuzzing
- add rt_soundsource3d_rebase_origin so floating-origin rebases shift unbound spatial sources by the world delta while node-bound sources continue to follow the scene rebase

Tests:
- new unit regression (test_rt_gltf_draco_probe) and libFuzzer harness (fuzz_gltf_draco) that drive the Draco decoder directly and assert malformed input is rejected without a trap or out-of-bounds access
- 2D physics coverage for distance-joint settling and shortest-axis ejection, 3D math-handle and inverse-inertia coverage, and updated SpriteBatch and Canvas contract expectations for the blended region path and full-precision gradients
- register the two new internal runtime symbols in the surface policy

Reference demos (game content only; no Viper source is changed by demo work):
- Neon Lanes (3dbowling): commercial front end, responsive 1280x720 and compact layouts, versioned save migration, a dressed venue with a release lighting and material pipeline, and a readable delivery model that replaces the oversized automatic-strike corridor, with new gameplay modules and a broad deterministic probe suite
- Ashfall: a combat-first FPS campaign in which all nine missions are won by fighting, surviving an assault, or moving through a combat-secured exit, with new enemy projectiles and environmental hazards and preserved headless and software-render coverage
- Ridgebound: a 20-point upgrade covering uneven-terrain locomotion, no-jump route-first traversal, a genuinely submerged lake basin, day/night and slow-weather cycles with deterministic reset, and grounded art direction, with new topology, traversal, and lifecycle probes

Tests: Not run per user request.

+++ src/runtime/graphics/2d/rt_pixels_png.c
@@ -1071,7 +1071,9 @@ int64_t rt_pixels_save_png(void *pixels_ptr, void *path) {
 
     for (uint32_t y = 0; y < h; y++) {
         for (uint32_t x = 0; x < w; x++) {
-            uint32_t pixel = p->data[y * w + x];
+            /* Index in size_t: y*w computed in uint32_t would wrap for an image with
+             * more than 2^32 pixels, reading the wrong source pixel. */
+            uint32_t pixel = p->data[(size_t)y * (size_t)w + x];
             cur_row[x * 4 + 0] = (uint8_t)((pixel >> 24) & 0xFF); // R
             cur_row[x * 4 + 1] = (uint8_t)((pixel >> 16) & 0xFF); // G
             cur_row[x * 4 + 2] = (uint8_t)((pixel >> 8) & 0xFF);  // B
